Change a bunch of code to allow multiple waypoints
Instead of a singel center and destination there is a place vector. The first
place is where the model is at now. Any furthur are future places, or
destinations/waypoints. The code drawing path has been updated for this change.
Once the model is within 10 of the destination it is removed. If it is within 25
and has a second waypoint, speed is double for smoother transitions